feat: implement consistent pause functionality in options page and abstract logic
